Support event bubbling across shadowroot.
Normally, ShadowRoots are isolated. For example, document.querySelector won't
search within a ShadowRoot. This works pretty well in libdom, since it has no
concept of a ShadowRoot - we simply never append the ShadowRoot to libdom's
DOM representation.

Except, that doesn't work for events, which don't share that same isolation
behavior. Clicking a <div> inside a ShadowRoot bubbles all the way up to the
Window - like any other event.

To support this, we add an optional host: *Node to DocumentFragements. The host
is the parent element of the shadowroot. If we're bubbling an event, and reach
a DocumentFragment which has a host, then we continue bubbling up the host.