When namespace is NULL, match attributes by name only
I believe this is the correct behavior for getAttribute, setAttribute and
removeAttribute - they should work on attributes by name only. So I made this
change deep, in `_dom_element_attr_list_find_by_name` to cover all cases. It
seems safe to do, but there could be a place that relies on strict namespace
null matching...